template-src-linux-driver: Fix condition for kernel.targetinstall.post
The condition for including the rule to make
$(STATEDIR)/kernel.targetinstall.post depend on
$(STATEDIR)/@package@.targetinstall is broken because it uses
PTX_PACKAGES_SELECTED, which is always empty at the time this condition
is evaluated.

If PTX_COLLECTION is 'y', the rule should be included if
PTXCONF_@PACKAGE@ is 'y'.  If PTX_COLLECTION is unset, the rule should
be included if PTXCONF_@PACKAGE@ is either 'y' or 'm'.

Reported-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>