at91bootstrap2: Make targetinstall verbose and fix message
The additional -vD is also used in several other rules, it's nice to see
the image(s) are actually copied to the right place.

If there's no kconfig file yet, the message shown is right, but the
additional single quote leads to an error, when directly copy/paste to
the commandline.

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>