u-boot: Add option to use Kconfig based configuration
Recent versions of U-Boot (from 2014.10) use a Kconfig based
configuration. To also support the old system and not break existing
builds, this is introduced as optional, which also allows to build a
defconfig without oldconfig/menuconfig and a .config file in the BSP.

A new menu entry of type 'choice' was added to choose between the new
Kconfig based and the legacy config system (boldly inspired by
buildroot). Options for prepare and compile stage were revised and
adapted to U-Boot build documentation (CROSS_COMPILE and HOSTCC in env
instead of make options).

That part is based on the first patch 'u-boot: add support for
oldconfig/menuconfig' by Ahmad Fatoum from October 2018, but extended by
the new menu options, which more or less follows a suggestion of Michael
Olbrich to support both configuration systems and distinguish in the
make rule.

The option to avoid parallel building is only kept for the legacy build,
modern U-Boot uses the kernel build system and should build fine in
parallel.

Also added is support for ptxdist option -v for a more verbose build.
This is passed to make with V= option, which is based on a second patch
'u-boot: add V=$(PTXDIST_VERBOSE) to make options' by Ahmad Fatoum, but
squashed into this, because all those variable stuff was rewritten
anyway.

Compile tested with:

* recent DistroKit, ptxdist 2019.01.0, U-Boot 2019.01, and
  defconfig for Microchip SAMA5D27-SOM1-EK1 (platform-v7a), both build
  system approaches.
* custom BSP, ptxdist 2018.05.0, U-Boot 2012.04.01, and config
  "at91sam9g20ek_nandflash" (arm-v5te), legacy build only.

Co-authored-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Alexander Dahl <ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>