doc: 'notice' environment was replaced by 'sphinxadmonition'
Spinx deprecated 'notice' environment in 1.6b1 and finally removed in
1.7.0b1.

Anyway, environment renewal to show warning symbol for note rst
directive was not working in between.

This patch both fixes warning symbol in sphinx versions between 1.6b1
and 1.7.0b1 and tex generation error after version 1.7.0b1:

| ! LaTeX Error: Environment notice undefined.

Instead of renewing a possible non-existing environment, we create a
new one and override sphinxadmonition environment only if it existed
before, or notice environment if not.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>