doc: remove search page and genindex from index
The ptxdist documentation does not use index directives (e.g.
".. index:: Foo"), so the automatically generated index page is always
empty. So remove it. If this directive is used some day we can add it
again.

The ptxdist theme already includes search functionality, so remove the
search directive as well.

These special names are documented here:

  https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#special-names

This is also a workaround for a sphinx bug accidentally showing these
directives in pdf output:

  https://github.com/sphinx-doc/sphinx/issues/2827

Signed-off-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>