doc: ref_make_macros: merge parameter macros into a single section
All of these macros can be explained concisely with one table and one
combined example, no need to repeat a whole section and paragraph of the
text for each one of them that only differ in a few words. Also mention
some use cases, and give real-life examples from existing code to better
explain the usage.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>