ptxd_kconfig_update_config: keep an empty diff for cosmetic changes
If comments or the order of symbols change then the diff is empty.
Create the diff anyways. Otherwise the sanity check in the prepare stage of
kconfig packages may fail.
This can happen for example with a new kernel patch stack that modifies the
kernel version in the top layer.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>