ptxd_kconfig: handle removed symbols when creating and applying diffs
If a dependency for a symbol is disabled, then the symbol will be dropped
from the config. In that case, it will not appear in the diff. This is not
exactly a problem, because kconfig will drop the symbol anyways. However,
mconf detects this as a modification and will always ask to save the
config.

To avoid this, add all dropped symbols in the diff and use that information
to drop them when applying the diff.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>