protobuf: add autogen.sh again
With the new version, the patch was applied upstream.
However building protoc still fails for architectures that need libatomic:
Libprotobuf uses libatomic but it is not found as an indirect dependency
with some toolchains.
After running autogen.sh -latomic is explicitly added to the linker options
and build protoc succeeds.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>