COPYING: make it clear that PTXdist is GPL-2.0-only, not -or-later
The existence of the GPLv2 license text in the repository implies that
PTXdist is licensed under at least version 2 of the license. However,
there are no hints as to whether the "or later" portion of the default
license dedication in the section "How to Apply These Terms to Your New
Programs" after the FSF GPL license text also applies to PTXdist. Make
clear that the "or later" portion does not apply, for the same reasons
as discussed in the Linux kernel project.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Acked-by: Robert Schwebel <r.schwebel@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>