Definitions: modify escaping $ in ptx/escape
Until now, $ was escaped with \\. This is not very useful, because the $ is
not actually escaped if the resulting string is passed to a shell.
Some packages add a extra \ to escape the $. However, this results in a \
in the shell that is not always acceptable.

Escape the $ with just one \ instead. All affected code in ptxdist is
modified as needed, but changes may be needed in BSP local rules.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>