opkg: move opkg-database to a read-only location
The directory tree in '/var/lib' might not be persistent and thus,
not in sync with the root filesystem. In this case it's required
to store the opkg database at a location which has a fixed
relation to the root filesystem's content it describes.

This change moves the opkg's database to a read-only location in
'/usr/share/opkg', which is then always in sync with the root
filesystem's content.

This changed opkg configuration works since commit
   a691341deb33077b9d5ede5fe349ee6b3fb99be1
 "ptxd_make_image_extract_xpkg_files: use ptxd_get_alternative to find opkg.conf"

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>