pre/barebox: only use <PKG>_VERSION to build the URL
When adding another barebox package to a BSP, e.g. named barebox-foo,
the BAREBOX_FOO variable will contain "barebox-foo-$(VERSION)". This
variable is used to build BAREBOX_DIR, BAREBOX_SOURCE, and
BAREBOX_FOO_URL.

We must to keep the customized package postfix in the build
directory, but in the latter case, the -foo postfix is transferred
into the URL, which will lead to errors while downloading the source.

We actually only need the barebox version to download the source
tarball, so use a static "barebox-" string to build the barebox/url
macro and also the BAREBOX_FOO_SOURCE variable in the package template,
so the same tarball isn't downloaded multiple times with different
names.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>