ptxdist: fix autoversion ptxconfig handling
The autoversion was changed to pick the version from the bottom layer. This
is accomplished by temporarily using the ptxconfig from that layer.
If the version is already correct, then no new ptxdist is executed.
As a result, the ptxconfig from the bottom layer is used. This fails with:

ptxdist: error: The selected config file:
ptxdist: error: BSP/base/configs/ptxconfig
ptxdist: error: is overwritten by:
ptxdist: error: BSP/configs/ptxconfig
ptxdist: error: The config file in the outer layer must be used!

Fix this by switching back to the correct ptxconfig after fetching the
version from the config.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>