ptxdist: enable clang with icecc
This is a bit more complicated. The symlink for clang that is passed to
icecc-create-env must contain a proper cross-compile prefix. Otherwise,
querying 'as' will return the wrong path (host instead of the toolchain
'as'). The same happens for 'objcopy'.
In addition icecc-create-env for all releases including 1.2 is broken and
resolves the symlink first, so the created environment contains the wrong
'as'.

So ptxdist needs a check for this and disable icecc clang support if
necessary.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>