inetd services: don't use busybox's find in postinst
busybox's find doesn't understand '-empty', so when running these
postinst scripts on a system without GNU coreutils, they will display a
nice help text instead.

Replace the find call with a test that tries to expand the same pattern
that is used afterwards, so cat doesn't run with a non-matching pattern.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>