rauc: select on-target tools depending on bootloader
RAUC relies on tools on the target to interact with the bootloader.
As the bootloader selection is not part of ptxconfig, but platformconfig,
have the user select the bootloader used on the target.
Based on this, auto-select the appropriate package:
  barebox: dt-utils (barebox-state)
  U-Boot: u-boot-tools (fw_printenv, fw_setenv)
  EFI: efibootmgr (efibootmgr)

Regarding GRUB support, grub-editenv from GRUB2 is required. Currently,
there is no "GRUB2" package in ptxdist, so no active option for it.

Signed-off-by: Andreas Pretzsch <apr@cn-eng.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>