ptxd_kconfig_create_config_merge: fix merging options that start with the symbol prefix
The 'config' script removes the CONFIG_ or PTXCONF_ prefix automatically if
an argument starts with it. Currently ptxd_kconfig_create_config_merge()
also removed the prefix when collecting the 'config' arguments.
This breaks if the actual symbol starts with the prefix. In this case, the
prefix is removed twice.
To fix this, pass the unmodified string and let 'config' remove the prefix
for all of them.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>