sqlite: version bump 3.19.2 -> 3.28.0
Clean up and pin down the config options. The removed defines in
CPPFLAGS are already set by the configure script when we supply the
right configure options; and this way ./scripts/configure_helper.py is
also happy. Add a patch so that building without pthreads does not fail
with linker errors (could already be broken in previous versions, but I
didn't test).

Even if "public_domain" is not a valid SPDX license identifier (see this
discussion [1] for background), it describes the license situation best.
Pin down the license MD5 mentioning that it is public domain.

[1]: https://wiki.spdx.org/view/Legal_Team/Decisions/Dealing_with_Public_Domain_within_SPDX_Files

Cc: Clemens Gruber <clemens.gruber@pqgruber.com>
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>