doc: remove empty chapter "Working with GIT sources"
All options for git:// <PKG>_URLs are explained in that same paragraph,
`git ptx-patches` is already explained in the "Patching Packages"
chapter, and out-of-tree building is explained in the Daily Work
chapter. If anything else related to Git is still open, this chapter can
be re-added later, but it is useless for now (and regularly confuses me
when looking for the location where all those things mentioned above are
explained).

Fixes: 185c2bebaa1686f2faac ("Split the daily-work section into separate files")
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>