ptxd_make_world_prepare/install/clean: sanitize sysroot
After a version bump or package configuration changes, files from the
previous version may still be in sysroot. This can cause all kinds of
problems.
Also, some packages fail to build if sysroot contains the files from the
same packages but with a different configuration.

To avoid this, do the following:
- remember <PKG>_PKGDIR in a way to find the old path after a version bump
- remove all files of a packages from sysroot at the beginning of the
  prepare stage and the install stage

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>