libinput: optionally install device quirks
Add an option LIBINPUT_QUIRKS to install the input device quirks data.

It is only useful for systems that allow connecting external input
devices, or for systems that are contained in the quirks database.
The size of the quirks database is a bit over 100 KiB.

If the input device quirk database is not installed, install a dummy
quirk file instead, to silence the libinput error message warning about
the missing quirk database.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
[mol: use menuconfig]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>