ptxd_make_world_get: handle packages that share the same source better
To allow multiple souce files for on packge, $(<source-file>) contains the
base variable to find the corresponding URL and md5sum. So in general this
needs to be used to call world/get and world/check_src.
However, when packages share the same source file then $(<source-file>)
points to just one of them. So others call world/get with the environment
for a different package.

This can have unexpected side effects. For example for kernel packages
created with the template that have the same version as the generic kernel
package: The special handling for the KERNEL_MAKEVARS -> KERNEL_MAKE_OPT
transition breaks the build of the new kernel package.

To avoid this, derive the environment from the current package if the
source file is <PKG>_SOURCE. In this case all necessary variables are
available. In all other cases stick to the current mechanism.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>