varoverlayfs: make run-varoverlayfs.mount a hidden dependency
The explict dependency would be nice because it allows systemd to find any
dependency loops.
However, we want to unmount /run/varoverlayfs at the end and the causes
problems if it is part of local-fs.target. With the latest systemd booting
just stops.
So mount /run/varoverlayfs manually. Also add 'set -e' to abort early
in case mounting fails.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>