image-root-ext: use mke2fs by default
Images created by genext2fs are a bit strange. For some reason, readdir()
returns d_type == DT_UNKNOWN. This breaks systemd unit parsing,
specifically alias (symlink) detection.
As a result. 'default.target' is detected as a 'normal' unit, not an alias
ans strange things happen. It looks like only 'basic.target' is started and
not 'multi-user.target' so gettys and other services are not started.

Just use mke2fs to create the image to avoid this.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>